+++ R/utils.R
@@ -392,13 +392,10 @@ treat_res <- function(res){
   averagingPeriod <- output$results$averagingPeriod
 
   if(!is.null(date)){
-
-    date <- dplyr::rename(
-      date,
-      date.utc = .data$utc,
-      date.local = .data$local
-      )
-
+    date <- dplyr::rename(date, date.utc = .data$utc)
+    if (!is.null(date[["local"]])) {
+      date <- dplyr::rename(date, date.local = .data$local)
+    }
   }
   if(!is.null(averagingPeriod)){
 
@@ -424,14 +421,14 @@ treat_res <- function(res){
   results <- dplyr::bind_cols(results, date)
   results <- dplyr::bind_cols(results, averagingPeriod)
 
-  results <- dplyr::tbl_df(results)
+  results <- dplyr::as_tibble(results)
 
 
   # get the meta
-  meta <- dplyr::tbl_df(
+  meta <- dplyr::as_tibble(
     as.data.frame(output$meta))
   #get the time stamps
-  timestamp <- dplyr::tbl_df(data.frame(
+  timestamp <- dplyr::as_tibble(data.frame(
     queriedAt = func_date_headers(res$response_headers$date)))
 
   attr(results, "meta") <- meta