+++ tests/testthat/test-aq_measurements.R
@@ -1,7 +1,7 @@
 test_that("The value_from and value_to arguments work as they should", {
 
   vcr::use_cassette("aq_measurements_Hyderabad_value_from", {
-   output <- aq_measurements(page = 1, city="Hyderabad", value_from=10, limit=10)
+    output <- aq_measurements(page = 1, city="Hyderabad", value_from=10, limit=10)
   })
 
 
@@ -27,10 +27,10 @@ test_that("measurements returns a data.frame (tbl_df)", {
   expect_that(attr(output, "meta"), is_a("tbl_df"))
 
   vcr::use_cassette("aq_measurements_misc", {
-  expect_that(aq_measurements(page = 1, attribution = TRUE, limit = 2), is_a("tbl_df"))
-  expect_that(aq_measurements(page = 1, source_name = TRUE, limit = 2), is_a("tbl_df"))
-  expect_that(aq_measurements(page = 1, averaging_period = TRUE, limit = 2), is_a("tbl_df"))
-  expect_that(aq_measurements(page = 1, averaging_period = TRUE, attribution = TRUE, limit = 2), is_a("tbl_df"))
+    expect_that(aq_measurements(page = 1, attribution = TRUE, limit = 2), is_a("tbl_df"))
+    expect_that(aq_measurements(page = 1, source_name = TRUE, limit = 2), is_a("tbl_df"))
+    expect_that(aq_measurements(page = 1, averaging_period = TRUE, limit = 2), is_a("tbl_df"))
+    expect_that(aq_measurements(page = 1, averaging_period = TRUE, attribution = TRUE, limit = 2), is_a("tbl_df"))
   })
 
 })
@@ -43,21 +43,21 @@ test_that("measurements has the right columns", {
                               city="Chennai")
   })
 
-  expect_true(class(output$location) == "character")
-  expect_true(class(output$parameter) == "character")
-  expect_true(class(output$value) == "numeric" |
-                class(output$value) == "integer")
-  expect_true(class(output$unit) == "character")
-  expect_true(class(output$country) == "character")
-  expect_true(class(output$city) == "character")
-  expect_true(class(output$cityURL) == "character")
-  expect_true(class(output$locationURL) == "character")
-  expect_true(class(output$dateUTC)[1] == "POSIXct")
-  expect_true(class(output$dateLocal)[1] == "POSIXct")
-  expect_true(class(output$longitude) == "numeric" |
-                class(output$longitude) == "logical")
-  expect_true(class(output$latitude) == "numeric" |
-                class(output$latitude) == "logical")
+  expect_type(output$location, "character")
+  expect_type(output$parameter, "character")
+  expect_true(is(output$value, "numeric") || is(output$value, "integer"))
+  expect_type(output$unit, "character")
+  expect_type(output$country, "character")
+  expect_type(output$city, "character")
+  expect_type(output$cityURL, "character")
+  expect_type(output$locationURL, "character")
+
+  expect_true(is(output$dateUTC, "POSIXct"))
+  expect_true(is(output$dateLocal, "POSIXct") || is.null(output$dateLocal))
+  expect_true(is(output$longitude, "numeric") ||
+                is(output$longitude, "logical"))
+  expect_true(is(output$latitude, "numeric") ||
+                is(output$latitude, "logical"))
   meta <- attr(output, "meta")
   expect_true(all(names(meta) == c("name", "license",
                                    "website", "page",
@@ -76,5 +76,7 @@ test_that("the max value of page is used", {
                          parameter = "no2",
                          limit = 1)
 
-   expect_equal(nrow(res), 100)
+  expect_equal(nrow(res), 100)
 })
+
+