+++ NEWS.md
@@ -1,5 +1,11 @@
 # ropenaq (development version)
 
+# ropenaq 0.3.0
+
+* Fix code using dplyr (@hadley, #57)
+
+* Fix tests that used class() == .
+
 # ropenaq 0.2.9
 
 * Skip a vcr-enabled test on CRAN because of a (known, under investigation) vcr problem with encoding. https://github.com/ropensci/vcr/issues/158